+++ patzilla/commands.py
@@ -2,6 +2,7 @@
 # (c) 2017 Andreas Motl <andreas@ip-tools.org>
 import logging
 from docopt import docopt
+from patzilla.config import get_configuration
 from patzilla.util.logging import boot_logging
 from patzilla.version import __version__
 
@@ -12,6 +13,7 @@ APP_NAME = 'patzilla'
 def run():
     """
     Usage:
+      {program} make-config <config-kind>
       {program} info
       {program} --version
       {program} (-h | --help)
@@ -39,3 +41,7 @@ def run():
     # Debugging
     #print('options: {}'.format(options))
 
+    if options['make-config']:
+        kind = options['<config-kind>']
+        payload = get_configuration(kind)
+        print(payload)